Change time style to long format
Adds seconds and timezone to formatted datetimes.

+++ src/lib/util.ts
@@ -7,14 +7,16 @@ export const formatJson = (maybeJson: Record<string, unknown> | undefined) : str
 export function formatDate(timestamp: number): string {
   const date: Date = new Date(+timestamp);
   const lang: string = navigator.language;
+
   const formatOptions: Intl.DateTimeFormatOptions = {
     dateStyle: 'long',
-    timeStyle: 'short'
+    timeStyle: 'long'
   };
 
   const formattedDate: string = new Intl.DateTimeFormat(
     lang,
     formatOptions
   ).format(date);
+
   return formattedDate;
 }
